refactor: 하객 폼 페이지 중복 코드 제거
